+++ src/client/components/test/context.js
@@ -13,6 +13,15 @@ const protocolRegEx = /^(?:(?:https?|ftp):\/\/)/i;
 const urlRegEx = /^(?:(?:https?|ftp):\/\/)?(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,}))\.?)(?::\d{2,5})?(?:[/?#]\S*)?$/; // eslint-disable-line
 const base64ImgRegEx = /^data:image\/([a-zA-Z0-9-_.])+;base64,([^"]*)$/i;
 
+const isVideo = str => {
+  if (!isString(str)) {
+    return false;
+  }
+
+  const hashIndex = str.indexOf('#');
+  return videoRegEx.test(hashIndex > 0 ? str.slice(0, hashIndex) : str);
+}
+
 class TestContext extends Component {
   static displayName = 'TestContext';
 
@@ -80,7 +89,7 @@ class TestContext extends Component {
 
   renderContextContent = (content, title, highlight = false) => {
     // Videos
-    if (videoRegEx.test(content)) {
+    if (isVideo(content)) {
       return this.renderVideo(content, title);
     }